Correct it so that the variable saved on the server-side is stored specifically for each citizenid.
Since a variable is being used that is saved on the server-side, if a single event from one player sets it to true, it remains set to true for all others. That’s why the first creation was working, but the following ones were not, because this validation should be individual and not global.